Redesigned UnitBlock
UnitBlock now has the following Variables:

- commitment
- power
- primary spinning reserve
- secondary spinning reserve
- heat

For now, all those variables are optional in the sense that they are
not generated unless required. When generating the abstract variables,
the variables that should be generated must be indicated using a
SimpleConfiguration<int>. The first bits of the int in this
Configuration indicate which Variables must be generated.