Redesigned NetworkBlock
A NetworkBlock is associated with a time in {1, ..., T}, where T is
the time horizon, and has a pointer to a Network. Besides some data of
the network (demand, etc.), it also has a list of flow limit
constraints.

The computation of the power transfer distribution factor matrix, that
is used to generate the flow limit constraints, is still to be done.